added link-based state sharing
